Compilation on Windows with Microsoft Visual Studio 2008
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@217 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ wxParams/src/wxNumericParameter.h
@@ -80,7 +80,7 @@ public:
   void SetDecimate( int d)  { 
     _decimate=d; 
     if (_text!=NULL)
-      _text->SetSize(wxSize( 10*(int)log10(_max)+
+      _text->SetSize(wxSize( 10*(int)log10((double)_max)+
                              10*_decimate,
                              wxDefaultCoord));
     _factor=1;
